# openapi.yaml
openapi: 3.1.0
info:
  title: Cecilia API
  version: "2.0.0"

servers:
  - url: https://chat-gdatabot.onrender.com

security:
  - BearerAuth: []

paths:
  /validations/document:
    post:
      operationId: validateDocument
      tags:
        - Validações
      summary: Valida CPF ou CNPJ
      description: Valida documentos CPF e CNPJ, incluindo formato, quantidade de caracteres e dígitos verificadores.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateDocumentRequest"
            examples:
              cpf:
                summary: Validação de CPF
                value:
                  tipo: cpf
                  valor: "12345678909"
              cnpj:
                summary: Validação de CNPJ
                value:
                  tipo: cnpj
                  valor: "12345678000195"
      responses:
        "200":
          description: Resultado da validação do documento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidateDocumentResponse"
              examples:
                validCpf:
                  summary: CPF válido
                  value:
                    ok: true
                    documento_original: "12345678909"
                    documento_valido: true
                    tipo_documento: CPF
                    status_documento: VALID_CPF
                    motivo_invalido: ""
                invalidCpf:
                  summary: CPF inválido
                  value:
                    ok: true
                    documento_original: "123.456.789-09"
                    documento_valido: false
                    tipo_documento: CPF
                    status_documento: INVALID_CPF_FORMAT
                    motivo_invalido: CPF deve conter apenas números, sem pontos, traços, letras ou espaços
        "403":
          description: Token inválido
        "422":
          description: Erro de validação do payload

  /adm/client-charge:
    patch:
      operationId: atualizarCobrancaCliente
      tags:
        - Sinal
      summary: Atualizar Cobrança de Cliente
      description: |
        Atualiza os campos de cobrança exibidos no módulo Kanban do Financeiro,
        localizando os registros por contato.
        O contato pode vir com +, espaços, hífens ou parênteses; a aplicação
        normaliza o valor para somente números antes da busca.

        Somente os campos abaixo podem ser alterados por este endpoint:
        - status
        - resposta
        - dt_projecao_pgto
        dt_cobranca é preenchida automaticamente pela aplicação com a data e hora
        atual de São Paulo sempre que a atualização é processada.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientChargeRequest"
            examples:
              botconversa:
                summary: Atualização enviada pelo BotConversa
                value:
                  contato: "+5521999999999"
                  status: Atual (Com Projeção)
                  resposta: Cliente informou previsão de pagamento.
                  dt_projecao_pgto: "2026-06-30"
      responses:
        "200":
          description: Cobrança atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateClientChargeResponse"
              examples:
                updated:
                  summary: Atualização realizada
                  value:
                    ok: true
                    updated_count: 1
                    contato: "5521999999999"
        "400":
          description: Nenhum campo de cobrança informado
        "403":
          description: Token inválido ou ausente
        "404":
          description: Nenhuma cobrança encontrada para o filtro informado
        "422":
          description: Erro de validação do payload
        "503":
          description: Falha ao atualizar a cobrança

  /chatbots/registrarAtendimento:
    post:
      operationId: registrarAtendimento
      tags:
        - Chatbots
      summary: Registrar Atendimento
      description: Registra um atendimento iniciado pela aplicação de chatbot.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatbotEventRequest"
            examples:
              atendimento:
                summary: Atendimento iniciado
                value:
                  id_cliente: "12345"
                  tags: NomeCadastrado
                  dt_interacao: "2026-06-26T14:30:00"
                  bot: CECILia
      responses:
        "200":
          description: Registro criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatbotRecordResponse"
        "400":
          description: Payload vazio
        "403":
          description: Token inválido ou ausente
        "422":
          description: Payload inválido
        "503":
          description: Falha ao registrar o evento

  /chatbots/registrarContatoInicial:
    post:
      operationId: registrarContatoInicial
      tags:
        - Chatbots
      summary: Registrar Contato Inicial
      description: Registra o primeiro contato do cliente na jornada do chatbot.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatbotEventRequest"
            examples:
              contatoInicial:
                summary: Contato inicial
                value:
                  id_cliente: "12345"
                  motivo: Financeiro
                  menu: Segunda via
                  dt_interacao: "2026-06-26T14:31:00"
                  bot: CECILia
      responses:
        "200":
          description: Registro criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatbotRecordResponse"
        "400":
          description: Payload vazio
        "403":
          description: Token inválido ou ausente
        "422":
          description: Payload inválido
        "503":
          description: Falha ao registrar o evento

  /chatbots/registrarInteracao:
    post:
      operationId: registrarInteracao
      tags:
        - Chatbots
      summary: Registrar Interação
      description: Registra uma interação ocorrida durante o atendimento no chatbot.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatbotEventRequest"
            examples:
              interacao:
                summary: Interação registrada
                value:
                  id_cliente: 12345
                  id_interacao: 1
                  resumo: Cliente solicitou segunda via do boleto.
                  dt_interacao: "2026-06-26T14:32:00"
                  bot: CECILia
      responses:
        "200":
          description: Registro criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatbotRecordResponse"
        "400":
          description: Payload vazio
        "403":
          description: Token inválido ou ausente
        "422":
          description: Payload inválido
        "503":
          description: Falha ao registrar o evento

  /chatbots/registrarAvaliacaoAtendimento:
    post:
      operationId: registrarAvaliacaoAtendimento
      tags:
        - Chatbots
      summary: Registrar Avaliação de Atendimento
      description: Registra a avaliação enviada pelo cliente ao final do atendimento.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatbotEventRequest"
            examples:
              avaliacao:
                summary: Avaliação registrada
                value:
                  id_cliente: 12345
                  nota: "5"
                  comentario: Atendimento rápido.
                  dt_interacao: "2026-06-26T14:40:00"
                  bot: CECILia
      responses:
        "200":
          description: Registro criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatbotRecordResponse"
        "400":
          description: Payload vazio
        "403":
          description: Token inválido ou ausente
        "422":
          description: Payload inválido
        "503":
          description: Falha ao registrar o evento

  /chatbots/registrarAtendimentoHumano:
    post:
      operationId: registrarAtendimentoHumano
      tags:
        - Chatbots
      summary: Registrar Atendimento Humano
      description: Registra o encaminhamento ou atendimento realizado por uma pessoa.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatbotEventRequest"
            examples:
              atendimentoHumano:
                summary: Atendimento humano
                value:
                  id_cliente: 12345
                  pr_resumo: Cliente pediu suporte humano.
                  tp_atendimento: Financeiro
                  dt_interacao: "2026-06-26T14:35:00"
                  bot: CECILia
      responses:
        "200":
          description: Registro criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatbotRecordResponse"
        "400":
          description: Payload vazio
        "403":
          description: Token inválido ou ausente
        "422":
          description: Payload inválido
        "503":
          description: Falha ao registrar o evento

  /chatbots/registrarContatoFinal:
    post:
      operationId: registrarContatoFinal
      tags:
        - Chatbots
      summary: Registrar Contato Final
      description: Registra o encerramento do contato e se a demanda foi resolvida.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatbotEventRequest"
            examples:
              contatoFinal:
                summary: Contato final
                value:
                  id_cliente: 12345
                  pr_resolvidosn: S
                  dt_interacao: "2026-06-26T14:45:00"
                  bot: CECILia
      responses:
        "200":
          description: Registro criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatbotRecordResponse"
        "400":
          description: Payload vazio
        "403":
          description: Token inválido ou ausente
        "422":
          description: Payload inválido
        "503":
          description: Falha ao registrar o evento

  /chatbots/registrarServico:
    post:
      operationId: registrarServico
      tags:
        - Chatbots
      summary: Registrar Serviço
      description: Registra o serviço solicitado ou executado durante o atendimento.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatbotEventRequest"
            examples:
              servico:
                summary: Serviço registrado
                value:
                  id_cliente: 12345
                  servico: Segunda via de boleto
                  dt_interacao: "2026-06-26T14:33:00"
                  bot: CECILia
      responses:
        "200":
          description: Registro criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatbotRecordResponse"
        "400":
          description: Payload vazio
        "403":
          description: Token inválido ou ausente
        "422":
          description: Payload inválido
        "503":
          description: Falha ao registrar o evento

  /health/db:
    get:
      operationId: checarBancoDados
      tags:
        - Health
      summary: Checar Banco de Dados
      description: Verifica se a API consegue abrir conexão com o banco SQL Server.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Conexão com o banco validada
        "403":
          description: Token inválido ou ausente
        "503":
          description: Falha ao conectar no banco de dados

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ValidateDocumentRequest:
      type: object
      required:
        - tipo
        - valor
      properties:
        tipo:
          type: string
          description: Tipo do documento a validar.
          enum:
            - cpf
            - cnpj
          example: cpf
        valor:
          type: string
          description: Valor do documento sem pontuação.
          example: "12345678909"

    ValidateDocumentResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        documento_original:
          type: string
          example: "12345678909"
        documento_valido:
          type: boolean
          example: true
        tipo_documento:
          type: string
          example: CPF
        status_documento:
          type: string
          example: VALID_CPF
          enum:
            - VALID_CPF
            - INVALID_CPF_EMPTY
            - INVALID_CPF_FORMAT
            - INVALID_CPF_LENGTH
            - INVALID_CPF_DIGITS
            - VALID_CNPJ
            - INVALID_CNPJ_EMPTY
            - INVALID_CNPJ_FORMAT
            - INVALID_CNPJ_LENGTH
            - INVALID_CNPJ_DIGITS
            - INVALID_DOCUMENT_TYPE
        motivo_invalido:
          type: string
          example: ""

    UpdateClientChargeRequest:
      type: object
      required:
        - contato
        - status
      properties:
        contato:
          type: string
          minLength: 1
          maxLength: 40
          description: Valor usado para localizar a cobrança; aceita + e pontuação, mas será normalizado para números.
          example: "+5521999999999"
        status:
          type: string
          minLength: 1
          maxLength: 255
          description: Novo status da cobrança.
          example: Atual (Com Projeção)
        resposta:
          anyOf:
            - type: string
              maxLength: 2000
            - type: "null"
          description: Resposta ou observação retornada pelo cliente.
          example: Cliente informou previsão de pagamento.
        dt_projecao_pgto:
          anyOf:
            - type: string
              format: date
            - type: "null"
          description: Data projetada de pagamento no formato YYYY-MM-DD.
          example: "2026-06-30"

    UpdateClientChargeResponse:
      type: object
      required:
        - ok
        - updated_count
        - contato
      properties:
        ok:
          type: boolean
          description: Indica se a atualização foi processada.
          example: true
        updated_count:
          type: integer
          description: Quantidade de cobranças atualizadas.
          example: 1
        contato:
          type: string
          description: Contato usado no filtro.
          example: "5521999999999"

    ChatbotEventRequest:
      type: object
      additionalProperties: true
      description: Dados do evento enviados pela aplicação de chatbot.
      example:
        id_cliente: 12345
        dt_interacao: "2026-06-26T14:30:00"
        bot: CECILia

    ChatbotRecordResponse:
      type: object
      required:
        - ok
        - inserted_count
        - event
      properties:
        ok:
          type: boolean
          description: Indica se o registro foi processado.
          example: true
        inserted_count:
          type: integer
          description: Quantidade de registros inseridos.
          example: 1
        event:
          type: string
          description: Evento registrado pela aplicação.
          example: Registrar Atendimento
        reference:
          anyOf:
            - type: string
            - type: "null"
          description: Referência de rastreamento retornada em caso de falha.